# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mount Trail Booking API",
    "description": "API for booking mount expedition trails with group functionality and payment processing",
    "version": "1.0.0",
    "contact": {
      "email": "admin@mounttrail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://mounttrail-backend.test",
      "description": "Mount Trail API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter token in format (Bearer <token>)"
      }
    },
    "schemas": {
...
      "Booking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "group_code": {
            "type": "string",
            "example": "ABCD-EFGH"
          },
          "mount_id": {
            "type": "integer",
            "example": 1
          },
          "trail_id": {
            "type": "integer",
            "example": 1
          },
          "guide_id": {
            "type": "integer",
            "example": 2
          },
          "booking_date": {
            "type": "string",
            "format": "date",
            "example": "2023-12-01"
          },
          "start_time": {
            "type": "string",
            "format": "time",
            "example": "06:00"
          },
          "max_participants": {
            "type": "integer",
            "example": 10
          },
          "current_participants": {
            "type": "integer",
            "example": 3
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "cancelled", "completed"],
            "example": "pending"
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          },
          "data": {
            "type": "object"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/bookings": {
      "get": {
        "tags": ["Bookings"],
        "summary": "Get user's bookings",
        "security": [{"sanctum": []}],
        "responses": {
          "200": {
            "description": "User bookings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {"$ref": "#/components/schemas/ApiResponse"},
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Booking"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Bookings"],
        "summary": "Create a new booking",
        "security": [{"sanctum": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["mount_id", "trail_id", "guide_id", "booking_date", "start_time", "max_participants"],
                "properties": {
                  "mount_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "trail_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "guide_id": {
                    "type": "integer",
                    "example": 2
                  },
                  "booking_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2023-12-01"
                  },
                  "start_time": {
                    "type": "string",
                    "format": "time",
                    "example": "06:00"
                  },
                  "max_participants": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 20,
                    "example": 10
                  },
                  "notes": {
                    "type": "string",
                    "example": "Please bring warm clothes"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {"$ref": "#/components/schemas/ApiResponse"},
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Booking"
                        },
                        "group_code": {
                          "type": "string",
                          "example": "ABCD-EFGH"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/join": {
      "post": {
        "tags": ["Bookings"],
        "summary": "Join a booking group",
        "security": [{"sanctum": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["group_code"],
                "properties": {
                  "group_code": {
                    "type": "string",
                    "pattern": "^[A-Z]{4}-[A-Z]{4}$",
                    "example": "ABCD-EFGH"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully joined the group",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {"$ref": "#/components/schemas/ApiResponse"},
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Booking"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}